% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colonADR_main.R
\name{apcadr}
\alias{apcadr}
\title{apcadr}
\usage{
apcadr(x)
}
\arguments{
\item{x}{Row of a dataset.}
}
\description{
apcadr
}
\note{
This function calculates adenoma per colonoscopy (APC), adenoma detection rate (ADR), polyp per colonoscopy (PPC) and polyp detection rate (PDR) from JED-style database.
You can find sample database in "tests/testthat/sample_CF_data.csv".
}
\examples{
data <- c("Adenoma|adenoma|diverticulum", "diverticulum|hemorrhoid", "unremarkable", "FAP", "advanced colon cancer|polyp", "polyp|adenoma")
apcadr(data)

# returns "APC:0.6, ADR:0.4, PPC:1, PDR:0.66667

}
